Update the way a validation error displays for a letter.
There is an outstanding question about what the page title should be. For the time being we left it as Letter. But we will take a proper look at this in another story.
Use govuk-warning-text for cancelled letters.
Removed the test for permanent-failures since that's not a valid status type for letters.