Add notifyCheckboxes macro
Variant of govukCheckboxes with extra options to
allow the following we need in our pages:
1. setting the HTML tags for the
  `.govuk-checkboxes`  and
  `.govuk-checkboxes__item` elements
2. adding custom classes to the
   `.govuk-checkboxes__item` element
3. checkboxes to be nested, through adding a
   `.children` property to each item in
   `params.items`

Options 1. and 3. allow us to markup the
checkboxes as a nested list, something we use on
the user permissions and manage folder pages.

Option 2. is for the templates page, to let us
hide all templates/folders not present at the
point in the hierarchy the page represents. It's
also needed by the live-search functionality,
which shows/hides items in the list based on how
well they match the search term.

The base govukCheckboxes code was taken from:

https://github.com/alphagov/govuk-frontend/blob/v2.13.0/src/components/checkboxes/template.njk

Some fixes were then applied for where the
Nunjucks code doesn't work in Jinja, specifically:
- string concatenation in Jinja uses the `~`, not
  `+`
- making `params.items` `params['items']` to avoid
  Jinja thinking you're using the `.items`
  dictionary method
- `true` in JS being `True` in Python and `false`
  `False`
- quoting all dictionary keys

These fixes were mainly taken from:

https://github.com/alphagov/govuk-frontend-jinja