Remove POST param from agreement test
This test sent an empty string to represent the 'who' field not having a value. This is represented by a group of radios in the page.

In browsers, radios that aren't selected result in that field not being included in the HTTP parameters. This change makes the test match that.

The expected validation message is also changed to the one you see in a browser. Turns out [the WTForms DataRequired validator](https://wtforms.readthedocs.io/en/stable/validators.html#wtforms.validators.DataRequired) produces a different error message for an empty string vs the value being missing ("This field is required." vs "Not a valid choice").