Log ID when creating Zendesk ticket
Members of the public are still getting through to our support channel,
and we don’t really understand where they’re coming from. We don’t have
a way of tying their support ticket to anything in our logs.

By logging the ID of the ticket when we created it we should be able to
work back to the request they made to create the ticket.

From there, we should be able to look at what other requests they made,
to understand the journey they took to get to our support channel.