Bump gds metrics to 0.2.2 to bring in bug fix
https://github.com/alphagov/gds_metrics_python/blob/master/CHANGELOG.md#021

This will mean that if an exception is thrown in any `before_request`
functions that run before the gds metrics `before_request` function,
then we will no longer throw an uncaught exception.

Note, this could still be the case (and has been seen in production) as
`init_app(application)` in `create_app` in `init.py` can indeed throw
exceptions, for example when it can't find the current service. We could
move our init app for gds metrics further up if we wanted to to further
counter this.