Add confirmation step when cancelling user invite
There was no flash confirmation notice when you invited someone to your team
and then cancelled the invitation.

Having a flash confirmation banner is a common pattern used on other pages,
for example when you click the link to delete a template, and one that needs
to be added here according to the accessibility report.