Change mixins-before-declarations Sass-lint rule
We write our Sass to be mobile-first, meaning we
declare the value for a property (for mobile) and
then add all other variations (for other
viewports) afterwards.

Because of this, we need mixins that produce media
queries to be able to follow declarations.

This extension to the rule is also hinted at as
good practice in the docs:

https://gitlab.skotty.io/forks/sass-lint/blob/7847511b61b3ca01a86a8870c00b5a7fca94bac5/docs/rules/mixins-before-declarations.md